+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Icons: Add the following components:
+  - `MicrophoneIconNext`
+  - `QuestionBubbleIconNext`
+  - `UsersFilledIconNext`
+
 ## [10.8.0] - 2022-05-10
 
 Feature: